+++ source/xtreme3d/lightfx.pas
@@ -1,0 +1,10 @@
+function LightFXCreate(obj: real): real; cdecl;
+var
+  fx: TGLLightFX;
+  objct: TGLBaseSceneObject;
+begin
+  objct := TGLBaseSceneObject(RealToPtr(obj));
+  fx := GetOrCreateLightFX(objct);
+  result := PtrToReal(fx);
+end;
+